Creating new endpoint for ContactInfo history
Test on local works just fine but there will be a problem in the gui.
When we update an Asset in the gui it updates asset first and then contactinfo.
This causes one or two seconds difference between asset and contacinfo uipdate time.
So, finding a contactinfo with a given asset update time will result error.